Library update

+++ ESP32/main/SS.c
@@ -78,7 +78,7 @@ settings
 #define port_mask(p) ((p)&63)
 #define BITFIELDS "-"
 #define PORT_INV 0x40
-    uint16_t logical_gpio = 0;  // Logical GPIO (from GPIO 48, covers NFC, keypad...)
+    uint32_t logical_gpio = 0;  // Logical GPIO (from GPIO 48 to 63 and -48 to -63, covers NFC, keypad, GPS, etc.)
 
 // External
 const char *port_check(int p, const char *module, int in)
